requirements.txt
discord.py
requests
python-dotenv
boto3